+++ src/main/java/com/aidredaline/backend/domain/runningsession/dto/StartSessionReq.java
@@ -1,8 +1,8 @@
 package com.aidredaline.backend.domain.runningsession.dto;
 
 public record StartSessionReq(
-        Long userId,
-        Long routeId,
+        Integer userId,
+        Integer routeId,
         double startLat,
         double startLng
 ) {}
