+++ src/main/java/com/aidredaline/backend/domain/runningsession/dto/StartSessionRes.java
@@ -3,7 +3,7 @@ package com.aidredaline.backend.domain.runningsession.dto;
 import java.time.Instant;
 
 public record StartSessionRes(
-        Long sessionId,
+        Integer sessionId,
         String status,
         Instant startTime
 ) {}
