+++ src/main/java/com/aidredaline/backend/domain/runningsession/repository/GpsTrackingPointRepository.java
@@ -4,6 +4,6 @@ import com.aidredaline.backend.domain.runningsession.entity.GpsTrackingPoint;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface GpsTrackingPointRepository extends JpaRepository<GpsTrackingPoint, Long> {
-    List<GpsTrackingPoint> findBySessionIdOrderByRecordedAtAsc(Long sessionId);
+public interface GpsTrackingPointRepository extends JpaRepository<GpsTrackingPoint, Integer> {
+    List<GpsTrackingPoint> findBySessionIdOrderByRecordedAtAsc(Integer sessionId);
 }
